small changes - movments to tools and renaming
