+++ src/common/utils.py
@@ -5,7 +5,6 @@ import sys
 from typing import Any, Callable, Dict, Iterable, List, Tuple, Union
 import uuid
 
-sys.path.insert(0, os.getenv("HOME"))
 sys.path.insert(1, os.getenv("NOVA_HOME"))
 
 import importlib
