+++ tools/analyzer_pair_wise_distances.py
@@ -14,8 +14,8 @@ from src.datasets.label_utils import get_unique_parts_from_labels, get_cell_line
 from src.datasets.dataset_config import DatasetConfig
 from src.analysis.analyzer import Analyzer
 from src.common.utils import get_if_exists
-from NOVA.src.datasets.label_utils import get_batches_from_labels, get_unique_parts_from_labels, get_markers_from_labels
-from src.analysis.analyzer_pairwise_dist_utils import filter_by_labels, compute_distances, extract_pairs
+from src.datasets.label_utils import get_batches_from_labels, get_unique_parts_from_labels, get_markers_from_labels
+from tools.analyzer_pairwise_dist_utils import filter_by_labels, compute_distances, extract_pairs
 
 
 class AnalyzerPairwiseDistances(Analyzer):
