+++ tools/plot_attn_score_config.py
